Refactor layout components and update import paths
