+++ src/app/layout.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from "next";
 import localFont from "next/font/local";
 import "./globals.css";
-import LayoutWrapper from "@/app/components/LayoutWrapper";
+import LayoutWrapper from "@/app/components/(layouts)/LayoutWrapper";
 
 const geistSans = localFont({
   src: "./fonts/GeistVF.woff",
